requirements.txt
# Add these to requirements.txt
# Audio playback
pyaudio>=0.2.11
sounddevice>=0.4.6

# Search and analysis
whoosh>=2.7.4  # For text indexing
scikit-learn>=1.3.0  # For analysis
matplotlib>=3.7.0  # For visualizations
wordcloud>=1.9.2

# Internationalization
babel>=2.12.1
pyqt6-tools>=6.4.0

# Plugin system
importlib-metadata>=6.8.0
pluggy>=1.3.0

# Advanced export
reportlab>=4.0.4  # PDF export
pandas>=2.0.0